Improve multiple things
I meant to put this into multiple commits, but I apparently forgot to do
that and now I don't feel like splitting them apart.

This allows for JavaScript code generation (which I did not add in a
generic way; it's just JavaScript or nothing).

Also, it allows for updating the Blockly workspace based on a change to
initialXml as long as updateWorkspaceBasedOnXml is true. I needed this
so that I could load user scripts in Blockly and update
react-blockly-component.